Squashed commit of the following:
commit c9d266f95a25550ea173c03ba84f9aff9e7f6345
Author: fortinb <fortinb.perso@gmail.com>
Date:   Tue Jun 23 04:23:35 2026 -0400

    finalize membership creation

commit ef073b1f35a6329ed6097d79780789bd0d0e4e52
Author: fortinb <fortinb.perso@gmail.com>
Date:   Wed Jun 17 04:11:33 2026 -0400

    Process order, take payment, Create Membership

commit 874c56699eecb0f670f5eacb898236d72edaa7cb
Author: fortinb <fortinb.perso@gmail.com>
Date:   Sun Jun 14 06:36:29 2026 -0400

    add Moneris

commit f8b36b19be6ebf550bb7991af641535989bed86e
Author: fortinb <fortinb.perso@gmail.com>
Date:   Thu Jun 4 04:01:40 2026 -0400

    Add payment services

commit 53499f8d4a2b7ad96b126e70dcf859422dcba303
Author: fortinb <fortinb.perso@gmail.com>
Date:   Thu Jun 4 03:12:07 2026 -0400

    Create finance module

commit a9fcb81e0e294c38595356d617e025dd21abf73d
Author: fortinb <fortinb.perso@gmail.com>
Date:   Mon Jun 1 03:30:04 2026 -0400

    rename billingaddress to address

commit 1ae2715a116d1dc415564425cc840fdae3d5ba51
Author: fortinb <fortinb.perso@gmail.com>
Date:   Sat May 30 02:59:48 2026 -0400

    checkout

commit d279a63c9f6cb2213e227846ff5fb329a7e7d4f1
Author: fortinb <fortinb.perso@gmail.com>
Date:   Sat May 23 04:37:37 2026 -0400

    manage order

commit 7a92bf5bfcc9e00782e2496ffe271c5668da1ff5
Author: fortinb <fortinb.perso@gmail.com>
Date:   Wed May 20 06:02:18 2026 -0400

    manage order

commit 65deea64a89f9b120dcb949c94980145824c41d0
Author: fortinb <fortinb.perso@gmail.com>
Date:   Fri May 15 04:16:36 2026 -0400

    Cleanup Dtos + Order basic implementation

commit a0416ffb52b1133accf287d7ae5f731371694c2b
Author: fortinb <fortinb.perso@gmail.com>
Date:   Wed May 6 04:55:43 2026 -0400

    Add clean architecture order components

commit 41b7728d9239cecc69657967748d36407884c225
Author: fortinb <fortinb.perso@gmail.com>
Date:   Wed May 6 04:32:29 2026 -0400

    add order + paiement entity

+++ proj/hypo.domain.services/src/main/java/com/iso/hypo/brand/application/usecase/impl/CoachServiceImpl.java
@@ -168,9 +168,11 @@ public class CoachServiceImpl implements CoachService {
 	@Transactional
 	public void deleteAllByBrandUuid(String brandUuid) throws CoachException {
 		try {
-			long deletedCount = coachRepository.deleteAllByBrandUuid(brandUuid, requestContext.getUsername());
-			
-			logger.info("Coach deleted for brand - brandUuid={} deletedCount={} ", brandUuid, deletedCount);
+			if (brandQueryService.assertDeleted(brandUuid)) {
+				long deletedCount = coachRepository.deleteAllByBrandUuid(brandUuid, requestContext.getUsername());
+				
+				logger.info("Coach deleted for brand - brandUuid={} deletedCount={} ", brandUuid, deletedCount);
+			}
 		} catch (Exception e) {
 			logger.error("Error - brandUuid={}", brandUuid, e);
