+++ proj/hypo.domain.services/src/main/java/com/iso/hypo/membership/application/usecase/impl/MemberServiceImpl.java
@@ -9,7 +9,6 @@ import java.util.UUID;
 import org.modelmapper.ModelMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,19 +16,22 @@ import org.springframework.util.Assert;
 
 import com.iso.hypo.common.application.context.RequestContext;
 import com.iso.hypo.common.application.event.enumeration.OperationEnum;
+import com.iso.hypo.common.application.security.RoleEnum;
 import com.iso.hypo.common.domain.model.Message;
 import com.iso.hypo.common.domain.model.enumeration.MessageSeverityEnum;
-import com.iso.hypo.common.domain.model.enumeration.RoleEnum;
+import com.iso.hypo.common.domain.model.location.Address;
 import com.iso.hypo.membership.application.dto.MemberDto;
 import com.iso.hypo.membership.application.event.MemberEvent;
 import com.iso.hypo.membership.application.exception.MemberException;
 import com.iso.hypo.membership.application.mapper.MemberDtoMapper;
 import com.iso.hypo.membership.application.port.BrandServicePort;
 import com.iso.hypo.membership.application.port.UserServicePort;
+import com.iso.hypo.membership.application.port.dto.BrandRef;
 import com.iso.hypo.membership.application.port.dto.UserRef;
 import com.iso.hypo.membership.application.usecase.MemberService;
 import com.iso.hypo.membership.domain.model.Member;
 import com.iso.hypo.membership.domain.repository.MemberRepository;
+import com.iso.hypo.sale.application.exception.OrderException;
 
 @Service
 public class MemberServiceImpl implements MemberService {
@@ -43,9 +45,6 @@ public class MemberServiceImpl implements MemberService {
 	private final UserServicePort userServicePort;
 
 	private final ApplicationEventPublisher eventPublisher;
-
-	@Value("${app.test.run:false}")
-	private boolean testRun;
 
 	private static final Logger logger = LoggerFactory.getLogger(MemberServiceImpl.class);
 
@@ -70,10 +69,7 @@ public class MemberServiceImpl implements MemberService {
 
 			Member member = memberMapper.toEntity(memberDto);
 
-			if (!brandServicePort.brandExists(member.getBrandUuid())) {
-				throw new MemberException(requestContext.getTrackingNumber(), MemberException.BRAND_NOT_FOUND,
-						"Brand not found");
-			}
+			BrandRef brand = resolveBrand(memberDto.getBrandUuid());
 
 			// Find Member
 			Optional<Member> existingMember = memberRepository.findByBrandUuidAndPersonEmailAndDeletedIsFalse(
@@ -100,18 +96,22 @@ public class MemberServiceImpl implements MemberService {
 				newUser.setRoles(List.of(RoleEnum.member));
 
 				user = Optional.of(userServicePort.create(newUser, password, member.getBrandUuid()));
-				
-			/*	Message message = new Message();
-				message.setCode(MemberException.USER_ALREADY_EXIST);
-				message.setDescription("Duplicate user");
-				message.setSeverity(MessageSeverityEnum.critical);
-				member.setMessages(List.of(message));
 
-				throw new MemberException(requestContext.getTrackingNumber(), MemberException.USER_ALREADY_EXIST,
-						"Duplicate user", memberMapper.toDto(member));
-						*/
+				/*
+				 * Message message = new Message();
+				 * message.setCode(MemberException.USER_ALREADY_EXIST);
+				 * message.setDescription("Duplicate user");
+				 * message.setSeverity(MessageSeverityEnum.critical);
+				 * member.setMessages(List.of(message));
+				 * 
+				 * throw new MemberException(requestContext.getTrackingNumber(),
+				 * MemberException.USER_ALREADY_EXIST, "Duplicate user",
+				 * memberMapper.toDto(member));
+				 */
 			}
-			
+
+			member = initializeAddress(member, brand);
+
 			// Create member
 			member.setUuid(UUID.randomUUID().toString());
 			member.setCreatedOn(Instant.now());
@@ -128,6 +128,26 @@ public class MemberServiceImpl implements MemberService {
 			}
 			throw new MemberException(requestContext.getTrackingNumber(), MemberException.CREATION_FAILED, e);
 		}
+	}
+
+	private Member initializeAddress(Member member, BrandRef brand) {
+		// Set member address country and state same as brand if not provided in request
+		String country = brand.getAddress() != null ? brand.getAddress().getCountry() : "CA";
+		String state = brand.getAddress() != null ? brand.getAddress().getState() : "QC";
+
+		if (member.getPerson().getAddress() == null) {
+			member.getPerson().setAddress(new Address());
+		}
+
+		if (member.getPerson().getAddress().getCountry() == null) {
+			member.getPerson().getAddress().setCountry(country);
+		}
+
+		if (member.getPerson().getAddress().getState() == null) {
+			member.getPerson().getAddress().setState(state);
+		}
+		
+		return member;
 	}
 
 	@Override
@@ -166,7 +186,9 @@ public class MemberServiceImpl implements MemberService {
 	@Transactional
 	public MemberDto activate(String brandUuid, String memberUuid) throws MemberException {
 		try {
-			Member entity = this.readByMemberUuid(brandUuid, memberUuid);
+			BrandRef brand = resolveBrand(brandUuid);
+
+			Member entity = this.readByMemberUuid(brand.getUuid(), memberUuid);
 			entity.activate(requestContext.getUsername());
 			memberRepository.save(entity);
 
@@ -185,7 +207,9 @@ public class MemberServiceImpl implements MemberService {
 	@Transactional
 	public MemberDto deactivate(String brandUuid, String memberUuid) throws MemberException {
 		try {
-			Member entity = this.readByMemberUuid(brandUuid, memberUuid);
+			BrandRef brand = resolveBrand(brandUuid);
+			
+			Member entity = this.readByMemberUuid(brand.getUuid(), memberUuid);
 			entity.deactivate(requestContext.getUsername());
 			memberRepository.save(entity);
 
@@ -204,7 +228,9 @@ public class MemberServiceImpl implements MemberService {
 	@Transactional
 	public void delete(String brandUuid, String memberUuid) throws MemberException {
 		try {
-			Member entity = this.readByMemberUuid(brandUuid, memberUuid);
+			BrandRef brand = resolveBrand(brandUuid);
+			
+			Member entity = this.readByMemberUuid(brand.getUuid(), memberUuid);
 			entity.delete(requestContext.getUsername());
 			memberRepository.save(entity);
 
@@ -222,9 +248,11 @@ public class MemberServiceImpl implements MemberService {
 	@Override
 	public void deleteAllByBrandUuid(String brandUuid) throws MemberException {
 		try {
-			long deletedCount = memberRepository.deleteAllByBrandUuid(brandUuid, requestContext.getUsername());
+			if (brandServicePort.brandDeleted(brandUuid)) {
+				long deletedCount = memberRepository.deleteAllByBrandUuid(brandUuid, requestContext.getUsername());
 
-			logger.info("Member deleted for brand - brandUuid={} deletedCount={} ", brandUuid, deletedCount);
+				logger.info("Member deleted for brand - brandUuid={} deletedCount={} ", brandUuid, deletedCount);	
+			}
 		} catch (Exception e) {
 			logger.error("Error - brandId={}", brandUuid, e);
 
@@ -236,9 +264,11 @@ public class MemberServiceImpl implements MemberService {
 		try {
 			Assert.notNull(memberDto, "memberDto must not be null");
 
+			BrandRef brand = resolveBrand(memberDto.getBrandUuid());
+			
 			Member member = memberMapper.toEntity(memberDto);
 
-			Member oldMember = this.readByMemberUuid(member.getBrandUuid(), member.getUuid());
+			Member oldMember = this.readByMemberUuid(brand.getUuid(), member.getUuid());
 
 			ModelMapper mapper = new ModelMapper();
 			mapper.getConfiguration().setSkipNullEnabled(skipNull).setCollectionsMergeEnabled(false);
@@ -262,7 +292,7 @@ public class MemberServiceImpl implements MemberService {
 					throw new MemberException(requestContext.getTrackingNumber(), MemberException.MEMBER_ALREADY_EXIST,
 							"Duplicate member", memberMapper.toDto(oldMember));
 				}
-				
+
 				Optional<UserRef> user = userServicePort.findByEmail(member.getPerson().getEmail());
 				if (user.isPresent()) {
 					Message message = new Message();
@@ -276,18 +306,21 @@ public class MemberServiceImpl implements MemberService {
 				}
 			}
 
-			UserRef user = new UserRef();
-			user.setUuid(oldMember.getUserUuid());
-			user.setEmail(oldMember.getPerson().getEmail());
-			user.setFirstname(oldMember.getPerson().getFirstname());
-			user.setLastname(oldMember.getPerson().getLastname());
+			if (oldMember.getUserUuid() != null) {
+				UserRef user = new UserRef();
+				user.setUuid(oldMember.getUserUuid());
+				user.setEmail(oldMember.getPerson().getEmail());
+				user.setFirstname(oldMember.getPerson().getFirstname());
+				user.setLastname(oldMember.getPerson().getLastname());
 
-			// Don't update user role from member update API,
-			// as it is not in scope and can be managed separately from user management API.
-			// So set null to avoid overriding existing roles.
-			user.setRoles(null);
-			userServicePort.patch(user);
-
+				// Don't update user role from member update API,
+				// as it is not in scope and can be managed separately from user management API.
+				// So set null to avoid overriding existing roles.
+				user.setRoles(null);
+				userServicePort.patch(user);
+			}
+			
+			oldMember = initializeAddress(oldMember, brand);
 			oldMember.setModifiedOn(Instant.now());
 			oldMember.setModifiedBy(requestContext.getUsername());
 
@@ -312,6 +345,12 @@ public class MemberServiceImpl implements MemberService {
 		}
 
 		return entity.get();
+	}
+	
+	private BrandRef resolveBrand(String brandUuid) throws OrderException {
+		return brandServicePort.find(brandUuid)
+				.orElseThrow(() -> new OrderException(requestContext.getTrackingNumber(), OrderException.BRAND_NOT_FOUND,
+				"Brand not found - brandUuid=" + brandUuid));
 	}
 
 }
